+++ app/src/main/java/woowacourse/payments/ui/cardlist/CardListScreen.kt
@@ -1,5 +1,11 @@
 package woowacourse.payments.ui.cardlist
 
+import android.app.Activity.RESULT_OK
+import android.content.Context
+import android.content.Intent
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.ActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
@@ -14,6 +20,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
@@ -24,17 +31,32 @@ import woowacourse.payments.domain.CardNumber
 import woowacourse.payments.domain.CardholderName
 import woowacourse.payments.domain.ExpirationDate
 import woowacourse.payments.domain.Passcode
+import woowacourse.payments.ui.ExtraKeys
+import woowacourse.payments.ui.addcard.AddCardActivity
 import woowacourse.payments.ui.common.PaymentCard
+import woowacourse.payments.ui.format.ExpirationDateFormat
+import woowacourse.payments.ui.getParcelableExtraCompat
 import woowacourse.payments.ui.model.CardUiModel
 import woowacourse.payments.ui.model.toUiModel
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 import java.time.YearMonth
 
 @Composable
-fun CardListScreen(
-    cards: SnapshotStateList<CardUiModel>,
-    navigateToAddCard: () -> Unit,
-) {
+fun CardListScreen(cards: SnapshotStateList<CardUiModel>) {
+    val context: Context = LocalContext.current
+    val launcher =
+        rememberLauncherForActivityResult(ActivityResultContracts.StartActivityForResult()) { result: ActivityResult ->
+            if (result.resultCode == RESULT_OK) {
+                result.data?.toCardOrNull()?.let { card: Card ->
+                    cards.add(card.toUiModel())
+                }
+            }
+        }
+
+    fun navigateToAddCard() {
+        launcher.launch(AddCardActivity.intent(context))
+    }
+
     AndroidpaymentsTheme {
         Scaffold(
             modifier = Modifier.fillMaxSize(),
@@ -72,31 +94,40 @@ fun CardListScreen(
     }
 }
 
+private fun Intent.toCardOrNull(): Card? =
+    runCatching {
+        getParcelableExtraCompat<CardUiModel>(ExtraKeys.CARD_KEY)?.let { card: CardUiModel ->
+            val yearMonth =
+                YearMonth.parse(card.expirationDate, ExpirationDateFormat.formatPattern)
+            Card(
+                CardNumber(card.cardNumber),
+                ExpirationDate(yearMonth),
+                CardholderName(card.cardholderName),
+                Passcode(card.passcode),
+            )
+        }
+    }.getOrNull()
+
 @Preview(showBackground = true, name = "카드 목록 (0개)")
 @Composable
 private fun CardListScreenWithNoCardsPreview() {
-    CardListScreen(
-        cards = remember { mutableStateListOf() },
-        navigateToAddCard = {},
-    )
+    CardListScreen(remember { mutableStateListOf() })
 }
 
 @Preview(showBackground = true, name = "카드 목록 (1개)")
 @Composable
 private fun CardListScreenWithOneCardPreview() {
     CardListScreen(
-        cards =
-            remember {
-                mutableStateListOf(
-                    Card(
-                        CardNumber("1234123412341234"),
-                        ExpirationDate(YearMonth.of(2034, 12)),
-                        CardholderName("디랙"),
-                        Passcode("1234"),
-                    ).toUiModel(),
-                )
-            },
-        navigateToAddCard = {},
+        remember {
+            mutableStateListOf(
+                Card(
+                    CardNumber("1234123412341234"),
+                    ExpirationDate(YearMonth.of(2034, 12)),
+                    CardholderName("디랙"),
+                    Passcode("1234"),
+                ).toUiModel(),
+            )
+        },
     )
 }
 
@@ -121,6 +152,5 @@ private fun CardListScreenWithTwoCardsPreview() {
                     ).toUiModel(),
                 )
             },
-        navigateToAddCard = {},
     )
 }
